+++ src/core/abstractions/useFBO/component.vue
